chore: Move `OperationLog` tests module to separate file
